DBZ　mapper.xml最新化＆それに伴うソースの修正

+++ DB/DBZ/standard/business/src/main/java/jp/co/ndensan/reams/db/dbz/business/core/KyotakuKeikakuJikoSakuseiBuilder.java
@@ -13,7 +13,6 @@ import jp.co.ndensan.reams.ur.urz.definition.message.UrSystemErrorMessages;
 import jp.co.ndensan.reams.uz.uza.lang.FlexibleDate;
 import jp.co.ndensan.reams.uz.uza.lang.FlexibleYearMonth;
 import jp.co.ndensan.reams.uz.uza.lang.RString;
-import jp.co.ndensan.reams.uz.uza.math.Decimal;
 
 /**
  * {@link KyotakuKeikakuJikoSakusei}の編集を行うビルダークラスです。
@@ -71,7 +70,7 @@ public class KyotakuKeikakuJikoSakuseiBuilder {
      * @param 履歴番号 履歴番号
      * @return {@link KyotakuKeikakuJikoSakuseiBuilder}
      */
-    public KyotakuKeikakuJikoSakuseiBuilder set履歴番号(Decimal 履歴番号) {
+    public KyotakuKeikakuJikoSakuseiBuilder set履歴番号(int 履歴番号) {
         requireNonNull(履歴番号, UrSystemErrorMessages.値がnull.getReplacedMessage("履歴番号"));
         entity.setRirekiNo(履歴番号);
         return this;
