+++ DB/DBZ/standard/business/src/main/java/jp/co/ndensan/reams/db/dbz/business/core/KyotakuKeikakuJikoSakuseiTankiNyushoRiyoNissuBuilder.java
@@ -69,7 +69,7 @@ public class KyotakuKeikakuJikoSakuseiTankiNyushoRiyoNissuBuilder {
      * @param 履歴番号 履歴番号
      * @return {@link KyotakuKeikakuJikoSakuseiTankiNyushoRiyoNissuBuilder}
      */
-    public KyotakuKeikakuJikoSakuseiTankiNyushoRiyoNissuBuilder set履歴番号(Decimal 履歴番号) {
+    public KyotakuKeikakuJikoSakuseiTankiNyushoRiyoNissuBuilder set履歴番号(int 履歴番号) {
         requireNonNull(履歴番号, UrSystemErrorMessages.値がnull.getReplacedMessage("履歴番号"));
         entity.setRirekiNo(履歴番号);
         return this;
