+++ DB/DBZ/standard/persistence/src/main/java/jp/co/ndensan/reams/db/dbz/persistence/db/relate/KyotakuKeikakuJikoSakuseiDac.java
@@ -50,7 +50,7 @@ public class KyotakuKeikakuJikoSakuseiDac implements IModifiable<DbT3007KyotakuK
     @Transaction
     public Optional<DbT3007KyotakuKeikakuJikoSakuseiEntity> selectByKey(HihokenshaNo 被保険者番号,
             FlexibleYearMonth 対象年月,
-            Decimal 履歴番号) {
+            int 履歴番号) {
 
         requireNonNull(被保険者番号, UrSystemErrorMessages.値がnull.getReplacedMessage("被保険者番号"));
         requireNonNull(対象年月, UrSystemErrorMessages.値がnull.getReplacedMessage("対象年月"));
@@ -89,7 +89,7 @@ public class KyotakuKeikakuJikoSakuseiDac implements IModifiable<DbT3007KyotakuK
     @Transaction
     public Optional<DbT3007KyotakuKeikakuJikoSakuseiEntity> select直近居宅給付計画自己作成(HihokenshaNo 被保険者番号,
             FlexibleYearMonth 対象年月,
-            Decimal 履歴番号) {
+            int 履歴番号) {
 
         requireNonNull(被保険者番号, UrSystemErrorMessages.値がnull.getReplacedMessage("被保険者番号"));
         requireNonNull(対象年月, UrSystemErrorMessages.値がnull.getReplacedMessage("対象年月"));
