+++ DB/DBZ/standard/service/src/main/java/jp/co/ndensan/reams/db/dbz/service/core/basic/KyotakuKeikakuJikoSakuseiGokeiManager.java
@@ -59,7 +59,7 @@ public class KyotakuKeikakuJikoSakuseiGokeiManager {
     public KyotakuKeikakuJikoSakuseiGokei get居宅給付計画自己作成合計(
             HihokenshaNo 被保険者番号,
             FlexibleYearMonth 対象年月,
-            Decimal 履歴番号,
+            int 履歴番号,
             RString 居宅サービス区分,
             JigyoshaNo サービス提供事業者番号,
             ServiceShuruiCode サービス種類コード) {
